requirements.txt
django>=1.11,<1.12
django-debug-toolbar==1.8
gunicorn==19.4.5
whitenoise==3.3.1
mysqlclient
mysql-connector-python
Faker
